backend-flask/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.2
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
flask-bcrypt==1.0.1
email-validator==2.1.0 